pdf_manipulator/requirements.txt
astroid==1.5.3
backports.weakref==1.0rc1
bleach==1.5.0
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.1.2
ez-setup==0.9
Flask==0.12.2
html5lib==0.9999999
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Markdown==2.6.8
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.1
olefile==0.44
pbr==3.1.1
pdfminer.six==20170720
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.3.0
pycryptodome==3.4.6
Pygments==2.2.0
pylint==1.7.2
pyocr==0.4.7
simplegeneric==0.8.1
six==1.10.0
stevedore==1.25.0
traitlets==4.3.2
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wcwidth==0.1.7
Werkzeug==0.12.2
wrapt==1.10.10
